requirements.txt
fonttools>=4.38.0
Pillow>=9.0.0
PyYAML>=6.0
